correction reinitialiser les degats des attaques qui ont fait un coup critique

+++ src/main/java/fr/adsa/interfaces/IAttaque.java
@@ -8,6 +8,6 @@ public interface IAttaque {
     public double getChanceToucher();
     public void setNom(String nom);
     public void setChanceToucher(double chanceToucher);
-    public void setDegats(int degats);
+    public void setDegats(double degats);
     public int getDegats();
 }
